refactor(grading): group built-in graders under grading/graders/
Move the grader implementations (builtinGraders, humanGrader, llmJudge) into
lib/optimize/grading/graders/ so all built-in graders are easy to find; the rest
of grading/ (baseGrader, types, scorecard, aggregate, agencyRunner, getPath)
stays as the shared grading toolkit. Mirrors the optimizers/ layout. Updated the
one external importer (optimize.ts) and the graders' relative imports.
